Refactored repositories
