+++ src/main/java/com/beerjournal/breweriana/persistence/ItemCrudRepository.java
@@ -1,5 +1,6 @@
-package com.beerjournal.breweriana.persistence.item;
+package com.beerjournal.breweriana.persistence;
 
+import com.beerjournal.breweriana.persistence.item.Item;
 import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
@@ -7,7 +8,7 @@ import java.util.Collection;
 import java.util.Optional;
 import java.util.Set;
 
-public interface ItemCrudRepository extends MongoRepository<Item, ObjectId> {
+interface ItemCrudRepository extends MongoRepository<Item, ObjectId> {
     Optional<Item> findOneByName(String name);
     Set<Item> findByNameNotIn(Collection itemsNames);
 }
