+++ src/main/java/com/beerjournal/breweriana/persistence/UserCollectionCrudRepository.java
@@ -1,12 +1,13 @@
-package com.beerjournal.breweriana.persistence.collection;
+package com.beerjournal.breweriana.persistence;
 
 
+import com.beerjournal.breweriana.persistence.collection.UserCollection;
 import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.Optional;
 
-public interface UserCollectionCrudRepository extends MongoRepository<UserCollection, ObjectId> {
+interface UserCollectionCrudRepository extends MongoRepository<UserCollection, ObjectId> {
     Optional<UserCollection> findOneById(ObjectId id);
     Optional<UserCollection> findOneByOwnerId(ObjectId userId);
 }
