--- src/main/java/com/beerjournal/breweriana/persistence/category/CategoryCrudRepository.java
@@ -1,10 +1,0 @@
-package com.beerjournal.breweriana.persistence.category;
-
-import org.bson.types.ObjectId;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-import java.util.Optional;
-
-public interface CategoryCrudRepository extends MongoRepository<Category, ObjectId> {
-    Optional<Category> findOneByName(String name);
-}
