+++ src/test/groovy/com/beerjournal/breweriana/persistence/item/ItemRepositoryTest.groovy
@@ -1,7 +1,9 @@
 package com.beerjournal.breweriana.persistence.item
 
 import com.beerjournal.breweriana.persistence.category.Category
-import com.beerjournal.breweriana.persistence.category.CategoryCrudRepository
+import com.beerjournal.breweriana.persistence.CategoryCrudRepository
+import com.beerjournal.breweriana.persistence.ItemCrudRepository
+import com.beerjournal.breweriana.persistence.ItemRepository
 import com.beerjournal.breweriana.utils.TestUtils
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
